requirements.txt
charset_normalizer
import_expression
python-dotenv
braceexpand
discord.py
frozenlist
multidict
aiosignal
tabulate
requests
asqlite @ git+https://github.com/Rapptz/asqlite
jishaku
aiohttp
urllib3
certifi
chardet
attrs
yarl
idna
pytz